Create TranslationWidget
