+++ DailyWidget/DailyWidget.swift
@@ -8,29 +8,3 @@
 import WidgetKit
 import SwiftUI
 import DailyLibrary
-
-struct TranslationWidget: Widget {
-    let kind: String = "DailyWidget"
-
-    var body: some WidgetConfiguration {
-        StaticConfiguration(kind: kind, provider: TranslationTimelineProvider()) { entry in
-            TranslationView(entry: entry)
-                .containerBackground(.fill.tertiary, for: .widget)
-        }
-        .configurationDisplayName("Random Word")
-        .description("Learn new words every day.")
-        .supportedFamilies([.systemSmall, .systemMedium])
-    }
-}
-
-#Preview(as: .systemSmall) {
-    TranslationWidget()
-} timeline: {
-    TranslationTimelineEntry(date: Date.now, translation: Translation(from: Language.English.hello, to: Language.Italian.hello))
-}
-
-#Preview(as: .systemMedium) {
-    TranslationWidget()
-} timeline: {
-    TranslationTimelineEntry(date: Date.now, translation: Translation(from: Language.English.hello, to: Language.Italian.hello))
-}
